+++ oldSrc/library/DIPS.Mobile.UI/Components/Buttons/Button.Properties.cs
@@ -1,8 +1,0 @@
-using Xamarin.Forms;
-
-namespace DIPS.Mobile.UI.Components.Buttons
-{
-    public partial class Button
-    {
-    }
-}