+++ src/app/Components/MainPage.cs
@@ -1,12 +1,6 @@
-using System.Net.Mime;
-using System.Reflection;
-using Components.ComponentsSamples;
 using Components.Resources.LocalizedStrings;
-using Components.ResourcesSamples;
-using DIPS.Mobile.UI.Resources.Icons;
+using DIPS.Mobile.UI.Components.ListItems;
 using DIPS.Mobile.UI.Resources.Sizes;
-using DIPS.Mobile.UI.Sizes.Sizes;
-using Button = DIPS.Mobile.UI.Components.Buttons.Button;
 
 namespace Components;
 
@@ -19,33 +13,32 @@ public class MainPage : DIPS.Mobile.UI.Components.Pages.ContentPage
         Title = LocalizedStrings.Components;
         Content = new DIPS.Mobile.UI.Components.Lists.CollectionView()
         {
-            ItemsSource = sampleTypes, ItemTemplate = new DataTemplate(() => new NavigateToSamplesButton(samples)),
+            ItemsSource = sampleTypes, ItemTemplate = new DataTemplate(() => new NavigateToSamplesItem(samples)),
         };
     }
 }
 
-public class NavigateToSamplesButton : Button
+public class NavigateToSamplesItem : NavigationListItem
 {
     private readonly List<Sample> m_samples;
     private SampleType m_sampleType;
 
-    public NavigateToSamplesButton(List<Sample> samples)
+    public NavigateToSamplesItem(List<Sample> samples)
     {
         m_samples = samples;
-        Margin = 5;
-        this.SetBinding(TextProperty, new Binding(){Path = ""});
+        this.SetBinding(TitleProperty, new Binding(){Path = ""});
         Command = new Command(TryNavigateToSamplesPage);
     }
 
     private void TryNavigateToSamplesPage()
     {
-        var samples = m_samples.Where(sample => sample.Type == m_sampleType).ToList();
+        var samples = m_samples.Where(sample => sample.Type == m_sampleType).ToList().OrderBy(sample => sample.Name);
         if (!samples.Any())
         {
             Shell.Current.DisplayAlert("No samples",
                 $"Theres no samples for {m_sampleType} yet.", "Ok");
         }
-        
+
         Shell.Current.Navigation.PushAsync((new SamplesPage(m_sampleType, samples)));
     }