+++ src/app/Components/MauiProgram.cs
@@ -31,7 +31,6 @@ public static class MauiProgram
 #if DEBUG
         builder.Logging.AddDebug();
 #endif
-
         return builder.Build();
     }
 }