+++ src/app/Components/Platforms/Android/MainActivity.cs
@@ -1,7 +1,5 @@
 ﻿using Android.App;
 using Android.Content.PM;
-using Android.OS;
-using DIPS.Mobile.UI.Components.FloatingActionButton.FloatingActionButtonMenu.Android;
 
 namespace Components;
 
@@ -12,7 +10,6 @@ public class MainActivity : MauiAppCompatActivity
 {
     public MainActivity()
     {
-        
     }
     
 }