+++ src/app/Components/REGISTER_YOUR_SAMPLES_HERE.cs
@@ -29,10 +29,9 @@ public static class REGISTER_YOUR_SAMPLES_HERE
             new(SampleType.Resources, LocalizedStrings.Colors, () => new ColorsSamples()),
             new(SampleType.Resources, LocalizedStrings.Sizes, () => new SizesSamples()),
             new(SampleType.Components, LocalizedStrings.ListItems, () => new ListItemsSamples()),
-            new(SampleType.Components, LocalizedStrings.Chips, () => new ChipsSamples()),
+            new(SampleType.Components, LocalizedStrings.Chip, () => new ChipsSamples()),
             new(SampleType.Resources, LocalizedStrings.Icons, () => new IconsSamples()),
             new(SampleType.Components, LocalizedStrings.FABs, () => new FloatingActionButtonSamples())
-
         };
     }
 }