+++ src/app/Components/Resources/LocalizedStrings/LocalizedStrings.Designer.cs
@@ -123,9 +123,9 @@ namespace Components.Resources.LocalizedStrings {
             }
         }
         
-        internal static string Chips {
+        internal static string Chip {
             get {
-                return ResourceManager.GetString("Chips", resourceCulture);
+                return ResourceManager.GetString("Chip", resourceCulture);
             }
         }
         
@@ -138,6 +138,132 @@ namespace Components.Resources.LocalizedStrings {
         internal static string FABs {
             get {
                 return ResourceManager.GetString("FABs", resourceCulture);
+            }
+        }
+        
+        internal static string New_version {
+            get {
+                return ResourceManager.GetString("New_version", resourceCulture);
+            }
+        }
+        
+        internal static string New_version_message {
+            get {
+                return ResourceManager.GetString("New_version_message", resourceCulture);
+            }
+        }
+        
+        internal static string Download {
+            get {
+                return ResourceManager.GetString("Download", resourceCulture);
+            }
+        }
+        
+        internal static string Cancel {
+            get {
+                return ResourceManager.GetString("Cancel", resourceCulture);
+            }
+        }
+        
+        internal static string Birthday {
+            get {
+                return ResourceManager.GetString("Birthday", resourceCulture);
+            }
+        }
+        
+        internal static string SelectPerson {
+            get {
+                return ResourceManager.GetString("SelectPerson", resourceCulture);
+            }
+        }
+        
+        internal static string Context_Menu_SingleAction {
+            get {
+                return ResourceManager.GetString("Context_Menu_SingleAction", resourceCulture);
+            }
+        }
+        
+        internal static string Context_Menu_TwoActions {
+            get {
+                return ResourceManager.GetString("Context_Menu_TwoActions", resourceCulture);
+            }
+        }
+        
+        internal static string Context_Menu_Choices_Multiple {
+            get {
+                return ResourceManager.GetString("Context_Menu_Choices_Multiple", resourceCulture);
+            }
+        }
+        
+        internal static string Context_Menu_Choices_Single {
+            get {
+                return ResourceManager.GetString("Context_Menu_Choices_Single", resourceCulture);
+            }
+        }
+        
+        internal static string Context_Menu_GroupedActions {
+            get {
+                return ResourceManager.GetString("Context_Menu_GroupedActions", resourceCulture);
+            }
+        }
+        
+        internal static string Context_Menu_Grouped_Choices_Single {
+            get {
+                return ResourceManager.GetString("Context_Menu_Grouped_Choices_Single", resourceCulture);
+            }
+        }
+        
+        internal static string Context_Menu_LongPressWithAction {
+            get {
+                return ResourceManager.GetString("Context_Menu_LongPressWithAction", resourceCulture);
+            }
+        }
+        
+        internal static string Action_Format {
+            get {
+                return ResourceManager.GetString("Action_Format", resourceCulture);
+            }
+        }
+        
+        internal static string Group_Format {
+            get {
+                return ResourceManager.GetString("Group_Format", resourceCulture);
+            }
+        }
+        
+        internal static string BottomSheet_SimpleBottomSheet {
+            get {
+                return ResourceManager.GetString("BottomSheet_SimpleBottomSheet", resourceCulture);
+            }
+        }
+        
+        internal static string SearchPage_PlaceholderText {
+            get {
+                return ResourceManager.GetString("SearchPage_PlaceholderText", resourceCulture);
+            }
+        }
+        
+        internal static string SearchPage_HintText {
+            get {
+                return ResourceManager.GetString("SearchPage_HintText", resourceCulture);
+            }
+        }
+        
+        internal static string SearchPage_NoResultText {
+            get {
+                return ResourceManager.GetString("SearchPage_NoResultText", resourceCulture);
+            }
+        }
+        
+        internal static string Option_Format {
+            get {
+                return ResourceManager.GetString("Option_Format", resourceCulture);
+            }
+        }
+        
+        internal static string BottomSheet_Open {
+            get {
+                return ResourceManager.GetString("BottomSheet_Open", resourceCulture);
             }
         }
     }