+++ src/library/DIPS.Mobile.UI/Components/Chips/Android/ChipHandler.cs
@@ -1,17 +1,11 @@
-using Android.Content.Res;
-using DIPS.Mobile.UI.Platforms.Android;
-using DIPS.Mobile.UI.Resources.Colors;
+using DIPS.Mobile.UI.Components.Chips.Android;
 using Microsoft.Maui.Handlers;
-using Microsoft.Maui.Platform;
-using Colors = DIPS.Mobile.UI.Resources.Colors.Colors;
-using TextAlignment = Android.Views.TextAlignment;
 
 // ReSharper disable once CheckNamespace
 namespace DIPS.Mobile.UI.Components.Chips;
 
 public partial class ChipHandler : ViewHandler<Chip, Google.Android.Material.Chip.Chip>
 {
-
     protected override Google.Android.Material.Chip.Chip CreatePlatformView()
     {
         return new Google.Android.Material.Chip.Chip(Context) {  };
@@ -22,8 +16,15 @@ public partial class ChipHandler : ViewHandler<Chip, Google.Android.Material.Chi
         base.ConnectHandler(platformView);
         
         platformView.SetDefaultChipAttributes();
+        platformView.Click += OnChipTapped;
     }
-    
+
+    protected override void DisconnectHandler(Google.Android.Material.Chip.Chip platformView)
+    {
+        base.DisconnectHandler(platformView);
+        platformView.Click -= OnChipTapped;
+    }
+
     private static partial void MapTitle(ChipHandler handler, Chip chip)
     {
         handler.PlatformView.Text = chip.Title;