+++ src/library/DIPS.Mobile.UI/Components/ContextMenus/iOS/ContextMenuHelper.cs
@@ -7,7 +7,7 @@ namespace DIPS.Mobile.UI.Components.ContextMenus.iOS;
  {
     internal static Dictionary<ContextMenuItem, UIMenuElement> CreateMenuItems(
         IEnumerable<ContextMenuItem> contextMenuItems,
-        ContextMenu contextMenu, ContextMenuGroup menuGroup = null)
+        ContextMenu contextMenu, Action callBackWhenItemTapped = null, ContextMenuGroup menuGroup = null)
     {
         var dict = new Dictionary<ContextMenuItem, UIMenuElement>();
         var items = contextMenuItems.ToArray();
@@ -17,13 +17,16 @@ namespace DIPS.Mobile.UI.Components.ContextMenus.iOS;
             if (contextMenuItem is ContextMenuGroup contextMenuGroup) //Recursively add menu items from a group
             {
                 contextMenuGroup.Parent = contextMenu;
-                //Inherit isCheckable context menu group group to all menu items in the group
+                
                 foreach (var c in contextMenuGroup.ItemsSource)
                 {
-                    c.IsCheckable = contextMenuGroup.IsCheckable;
+                    if (!c.IsCheckable && contextMenuGroup.IsCheckable)//Inherit isCheckable from the parent context menu group if the item is not checkable
+                    {
+                        c.IsCheckable = true;
+                    }
                 }
 
-                var newDict = CreateMenuItems(contextMenuGroup.ItemsSource, contextMenu, contextMenuGroup);
+                var newDict = CreateMenuItems(contextMenuGroup.ItemsSource, contextMenu, callBackWhenItemTapped, contextMenuGroup);
                 if (items.Count(i => i is ContextMenuGroup) >
                     1) //If there is more than one group, add the group title and group the items
                 {
@@ -56,7 +59,7 @@ namespace DIPS.Mobile.UI.Components.ContextMenus.iOS;
                 }
                 
                 var uiAction = UIAction.Create(contextMenuItem.Title, image, null,
-                    uiAction => OnMenuItemClick(uiAction, contextMenuItem, contextMenu));
+                    uiAction => OnMenuItemClick(uiAction, contextMenuItem, contextMenu, callBackWhenItemTapped));
 
                 if (contextMenuItem.IsChecked)
                 {
@@ -83,23 +86,32 @@ namespace DIPS.Mobile.UI.Components.ContextMenus.iOS;
         return dict;
     }
 
-    private static void OnMenuItemClick(UIAction action, ContextMenuItem contextMenuItem,
-        ContextMenu contextMenu)
+    private static void OnMenuItemClick(UIAction action, ContextMenuItem tappedContextMenuItem,
+        ContextMenu contextMenu, Action? callBackWhenItemTapped)
     {
         
-        if (contextMenuItem.IsCheckable)
+        if (tappedContextMenuItem.IsCheckable)
         {
-            if (contextMenuItem.Parent is not ContextMenuGroup || !contextMenuItem.IsChecked)
-            { //Only check if if unchecked if its a part of a group
-                contextMenu.ResetIsCheckedForTheRest(contextMenuItem);
-
-                contextMenuItem.IsChecked =
-                    !contextMenuItem
-                        .IsChecked; //Can not change the visuals when the menu is showing as the items are immutable when they are showing    
+            var singleCheckMode = tappedContextMenuItem.Parent is ContextMenuGroup {IsCheckable: true};
+            
+            switch (singleCheckMode)
+            {
+                //You are unchecking an checked item that when single check mode is active, do not uncheck
+                case true when tappedContextMenuItem.IsChecked:
+                    return;
+                //You are checking an item that is not checked, reset the others
+                case true:
+                    contextMenu.ResetIsCheckedForTheRest(tappedContextMenuItem);
+                    break;
             }
+            
+            tappedContextMenuItem.IsChecked =
+                !tappedContextMenuItem
+                    .IsChecked; //Can not change the visuals when the menu is showing as the items are immutable when they are showing
         }
 
-        contextMenuItem.SendClicked(contextMenu);
+        tappedContextMenuItem.SendClicked(contextMenu);
+        callBackWhenItemTapped?.Invoke();
     }
 
     private static void SetCorrectUiActionState(ContextMenuItem contextMenuItem, UIAction uiAction)