+++ src/library/DIPS.Mobile.UI/Components/ListItems/ListItem.Properties.cs
@@ -1,4 +1,4 @@
-using DIPS.Mobile.UI.Sizes.Sizes;
+using DIPS.Mobile.UI.Resources.Sizes;
 
 namespace DIPS.Mobile.UI.Components.ListItems;
 
@@ -37,7 +37,7 @@ public partial class ListItem
         nameof(CornerRadius),
         typeof(CornerRadius),
         typeof(ListItem),
-        defaultValue: new CornerRadius(UI.Resources.Sizes.Sizes.GetSize(SizeName.size_4)),
+        defaultValue: new CornerRadius(Sizes.GetSize(SizeName.size_4)),
         propertyChanged: CornerRadiusChanged);
 
     /// <summary>