+++ src/library/DIPS.Mobile.UI/Components/Pickers/ItemPicker/ItemPickerBottomSheet.cs
@@ -1,7 +1,5 @@
 using System.Collections.ObjectModel;
 using DIPS.Mobile.UI.Components.BottomSheets;
-using DIPS.Mobile.UI.Extensions;
-using DIPS.Mobile.UI.Sizes.Sizes;
 using CheckBox = DIPS.Mobile.UI.Components.CheckBoxes.CheckBox;
 using SearchBar = DIPS.Mobile.UI.Components.Searching.SearchBar;
 using CollectionView = DIPS.Mobile.UI.Components.Lists.CollectionView;
@@ -11,10 +9,8 @@ namespace DIPS.Mobile.UI.Components.Pickers.ItemPicker
     public class PickerBottomSheet : BottomSheet
     {
         private readonly ItemPicker m_itemPicker;
-        private List<SelectableListItem> m_items;
         private readonly List<SelectableListItem> m_originalItems;
         private readonly SearchBar? m_searchBar;
-        private readonly CollectionView m_collectionView;
 
         public PickerBottomSheet(ItemPicker itemPicker)
         {
@@ -31,23 +27,23 @@ namespace DIPS.Mobile.UI.Components.Pickers.ItemPicker
 
             Items = new ObservableCollection<SelectableListItem>(m_originalItems);
 
-            var grid = new Grid() {RowDefinitions = new() {new(GridLength.Auto), new(GridLength.Star)}};
+            var grid = new Grid() {RowDefinitions = new RowDefinitionCollection {new(GridLength.Auto), new(GridLength.Star)}};
 
-            m_collectionView = new CollectionView()
+            var collectionView = new CollectionView()
             {
                 ItemTemplate = new DataTemplate(CreateCheckBox),
-                Margin = UI.Resources.Sizes.Sizes.GetSize(SizeName.size_2)
+                Margin = Sizes.GetSize(SizeName.size_2)
             };
 
             if (m_itemPicker.HasSearchBar)
             {
-                m_searchBar = new SearchBar() {HasCancelButton = false, BackgroundColor = Colors.Transparent};
+                m_searchBar = new SearchBar() {HasCancelButton = false, BackgroundColor = Microsoft.Maui.Graphics.Colors.Transparent};
                 grid.Add(m_searchBar, 0, 0);
             }
 
-            m_collectionView.SetBinding(ItemsView.ItemsSourceProperty, new Binding(nameof(Items), source: this));
+            collectionView.SetBinding(ItemsView.ItemsSourceProperty, new Binding(nameof(Items), source: this));
 
-            grid.Add(m_collectionView, 0, (m_searchBar != null) ? 1 : 0);
+            grid.Add(collectionView, 0, (m_searchBar != null) ? 1 : 0);
             Content = grid;
             SubscribeToEvents();
         }